+++ go/caddy/dal/models/admin_api_audit_log_models.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/inverse-inc/packetfence/go/caddy/pfpki/sql"
+	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/sql"
 	"gorm.io/gorm"
 )
