+++ go/caddy/dal/models/wrix.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/inverse-inc/packetfence/go/caddy/pfpki/sql"
+	"github.com/inverse-inc/packetfence/go/plugin/caddy2/pfpki/sql"
 	"gorm.io/gorm"
 
 	"strings"
